:art: Oppdatert mal med endringer
