+++ @navikt/core/css/primitives/page.css
@@ -43,7 +43,7 @@
   padding-inline: var(--a-spacing-4);
 }
 
-@media (min-width: 768px) {
+@media (min-width: 1024px) {
   .navds-pageblock--gutters {
     padding-inline: var(--a-spacing-12);
   }
